alot of tweaking!
-chances windows now allow you to put in decimels

-god trait stats and chances have been chanced, so now all of the gods stats are very similiar, with power in this order: (weakest to strongest)
god of night -> god of light _> god of knowledge -> god of chaos _> god of lich -> god of the earth -> god of war -> god of gods

also, can i give myself credit in mod.json?

+++ Code/GodChancesWindow.cs
@@ -53,7 +53,7 @@ namespace GodsAndPantheons
                         contents,
                         kv.Value.value
                     );
-                    input.inputField.characterValidation = InputField.CharacterValidation.Integer;
+                    input.inputField.characterValidation = InputField.CharacterValidation.Decimal;
                     input.inputField.onValueChanged.AddListener(delegate {
                         string pValue = NewUI.checkStatInput(input);
                         Main.modifyGodOption(ID, kv.Key, PowerButtons.GetToggleValue($"{kv.Key}Button"), pValue);
